.history/requirements_20251021114157.txt
fastapi
"uvicorn[standard]
sqlalchemy
psycopg2-binary
faker
python-dotenv
pydantic
